chore: move bootstrapFleet method to the test file
