+++ e2e/_suites/fleet/fleet_test.go
@@ -17,10 +17,12 @@ import (
 	"github.com/cenkalti/backoff/v4"
 	"github.com/cucumber/godog"
 	"github.com/cucumber/godog/colors"
+	"github.com/docker/go-connections/nat"
 	"github.com/elastic/e2e-testing/cli/config"
 	apme2e "github.com/elastic/e2e-testing/internal"
 	"github.com/elastic/e2e-testing/internal/common"
 	"github.com/elastic/e2e-testing/internal/deploy"
+	"github.com/elastic/e2e-testing/internal/elasticsearch"
 	"github.com/elastic/e2e-testing/internal/installer"
 	"github.com/elastic/e2e-testing/internal/kibana"
 	"github.com/elastic/e2e-testing/internal/shell"
@@ -215,6 +217,129 @@ func beforeScenario(fts *FleetTestSuite) {
 
 	fts.CurrentToken = enrollmentKey.APIKey
 	fts.CurrentTokenID = enrollmentKey.ID
+}
+
+// bootstrapFleet this method creates the runtime dependencies for the Fleet test suite, being of special
+// interest kibana profile passed as part of the environment variables to bootstrap the dependencies.
+func bootstrapFleet(ctx context.Context, env map[string]string) error {
+	deployer := deploy.New(common.Provider)
+
+	if profile, ok := env["kibanaProfile"]; ok {
+		log.Infof("Running kibana with %s profile", profile)
+	}
+
+	// the runtime dependencies must be started only in non-remote executions
+	return deployer.Bootstrap(ctx, deploy.NewServiceRequest(common.FleetProfileName), env, func() error {
+		kibanaClient, err := kibana.NewClient()
+		if err != nil {
+			log.WithFields(log.Fields{
+				"error": err,
+				"env":   env,
+			}).Fatal("Unable to create kibana client")
+		}
+
+		err = elasticsearch.WaitForClusterHealth(ctx)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"error": err,
+			}).Fatal("Elasticsearch Cluster is not healthy")
+		}
+
+		err = kibanaClient.RecreateFleet(ctx)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"error": err,
+				"env":   env,
+			}).Fatal("Fleet could not be recreated")
+		}
+
+		fleetServicePolicy := kibana.FleetServicePolicy
+
+		log.WithFields(log.Fields{
+			"id":          fleetServicePolicy.ID,
+			"name":        fleetServicePolicy.Name,
+			"description": fleetServicePolicy.Description,
+		}).Info("Fleet Server Policy retrieved")
+
+		maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute * 2
+		exp := utils.GetExponentialBackOff(maxTimeout)
+		retryCount := 1
+
+		fleetServerBootstrapFn := func() error {
+			serviceToken, err := elasticsearch.GetAPIToken(ctx)
+			if err != nil {
+				retryCount++
+				log.WithFields(log.Fields{
+					"error":       err,
+					"retries":     retryCount,
+					"elapsedTime": exp.GetElapsedTime(),
+				}).Warn("Could not get API Token from Elasticsearch.")
+				return err
+			}
+
+			fleetServerEnv := make(map[string]string)
+			for k, v := range env {
+				fleetServerEnv[k] = v
+			}
+
+			fleetServerPort, err := nat.NewPort("tcp", "8220")
+			if err != nil {
+				retryCount++
+				log.WithFields(log.Fields{
+					"error":       err,
+					"retries":     retryCount,
+					"elapsedTime": exp.GetElapsedTime(),
+				}).Warn("Could not create TCP port for fleet-server")
+				return err
+			}
+
+			fleetServerEnv["elasticAgentTag"] = common.ElasticAgentVersion
+			fleetServerEnv["fleetServerMode"] = "1"
+			fleetServerEnv["fleetServerPort"] = fleetServerPort.Port()
+			fleetServerEnv["fleetInsecure"] = "1"
+			fleetServerEnv["fleetServerServiceToken"] = serviceToken.AccessToken
+			fleetServerEnv["fleetServerPolicyId"] = fleetServicePolicy.ID
+
+			fleetServerSrv := deploy.ServiceRequest{
+				Name:    common.ElasticAgentServiceName,
+				Flavour: "fleet-server",
+			}
+
+			err = deployer.Add(ctx, deploy.NewServiceRequest(common.FleetProfileName), []deploy.ServiceRequest{fleetServerSrv}, fleetServerEnv)
+			if err != nil {
+				retryCount++
+				log.WithFields(log.Fields{
+					"error":       err,
+					"retries":     retryCount,
+					"elapsedTime": exp.GetElapsedTime(),
+					"env":         fleetServerEnv,
+				}).Warn("Fleet Server could not be started. Retrying")
+				return err
+			}
+
+			log.WithFields(log.Fields{
+				"retries":     retryCount,
+				"elapsedTime": exp.GetElapsedTime(),
+			}).Info("Fleet Server was started")
+			return nil
+		}
+
+		err = backoff.Retry(fleetServerBootstrapFn, exp)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"error": err,
+			}).Fatal("Fleet Server could not be started")
+		}
+
+		err = kibanaClient.WaitForFleet(ctx)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"error": err,
+				"env":   env,
+			}).Fatal("Fleet could not be initialized")
+		}
+		return nil
+	})
 }
 
 func setUpSuite() {
